refactor(rust): extract `try_rewrite_import_expression` logic into a standalone function
